requirements.txt
Flask==2.3.2
Pillow==12.0.0
